Move actions to actions/
