+++ src/onboarding/CameraRollContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import CameraRoll from 'src/profile/CameraRoll';
-import { changeProfileImage } from 'src/onboarding/actions';
+import { changeProfileImage } from 'src/actions/onboarding';
 
 export default connect(null, {
   onPhotoPress: path => changeProfileImage(path),
