+++ src/onboarding/ProfilePictureContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import ProfilePicture from 'src/profile/ProfilePicture';
-import { pushCameraRollRoute } from 'src/navigation/actions';
+import { pushCameraRollRoute } from 'src/actions/navigation';
 
 export default connect(
   state => ({
