+++ src/onboarding/WelcomeContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import Hello from './Hello';
-import { pushProfilePictureRoute } from 'src/navigation/actions';
+import { pushProfilePictureRoute } from 'src/actions/navigation';
 
 import { getProfileId } from 'src/reducers/auth';
 import { getProfile } from 'src/reducers/profile';
