+++ src/reducers/__tests__/auth-test.js
@@ -1,7 +1,7 @@
 import reduce from 'src/reducers/auth';
 import { fakeAuth, fakeProfileListData } from 'app/scripts/helpers';
 import * as statuses from 'src/auth/statuses';
-import * as actions from 'src/auth/actions';
+import * as actions from 'src/actions/auth';
 
 
 describe('auth/reducer', () => {
